Update README file; polishing test html pages (removed dependency on macropus)
