+++ test/test-xmllint.html
@@ -2,17 +2,17 @@
 <html>
   <head>
     <meta charset="utf-8">
-    <title>xpath_locator test</title>
+    <title>xmllint dtd test</title>
     <link rel="stylesheet" href="test.css">
   </head>
   <body>
-    <h1>xpath_locator test</h1>
+    <h1>xmllint dtd test</h1>
     <div id="status">Loading…</div>
 
     <pre id="lint-container"><code id="lint"></code></pre>
     <pre id="xml-container"><code id="xml"></code></pre>
 
     <script src="../xmltool.js"></script>
-    <script src="test-xpath-locator.js"></script>
+    <script src="test-xmllint.js"></script>
   </body>
 </html>
