+++ test/test-xmllint.xml
@@ -1,6 +1,12 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE article PUBLIC "-//NLM//DTD JATS (Z39.96) Journal Publishing DTD v1.0 20120330//EN" "JATS-journalpublishing1.dtd">
-<article xmlns:xlink="http://www.w3.org/1999/xlink" xmlns:mml="http://www.w3.org/1998/Math/MathML" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" article-type="research-article" dtd-version="1.0">
+<!DOCTYPE article 
+  PUBLIC "-//NLM//DTD JATS (Z39.96) Journal Publishing DTD v1.0 20120330//EN" 
+  "JATS-journalpublishing1.dtd">
+<article xmlns:xlink="http://www.w3.org/1999/xlink" 
+         xmlns:mml="http://www.w3.org/1998/Math/MathML" 
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" 
+         article-type="research-article" 
+         dtd-version="1.0">
   <front>
     <journal-meta>
       <journal-id journal-id-type="pmc">PeerJ</journal-id>
