--- test/test.html
@@ -1,14 +1,0 @@
-<!doctype html>
-<meta charset="utf-8">
-<title>xmllint dtd test</title>
-<link rel="stylesheet" href="test.css">
-
-<body>
-
-<div id="status">Loading…</div>
-
-<pre id="lint-container"><code id="lint"></code></pre>
-<pre id="xml-container"><code id="xml"></code></pre>
-
-<script src="../xmltool.js"></script>
-<script src="test.js"></script>
